requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
attrs==25.1.0
click==8.1.8
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.1
pyfiglet==1.0.2
python-dotenv==1.0.1
yarl==1.18.3
PyAudio==0.2.14
